pyproject.toml
[tool.poetry]
name = "pydanticGraph"
version = "0.1.0"
description = ""
authors = ["Nikita <n.mashalov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.3"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
networkx = "^3.2.1"
typer = "^0.9.0"
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"
hydra-core = "^1.3.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.2"

[tool.poetry.scripts]
start = 'pyvisgraph.__main__:app'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
